requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
black==23.9.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
